refactor(order): orderservice - kafka 데이터 누락 수정 - 빌드 에러 수정
