+++ api/admin-api/build.gradle
@@ -6,7 +6,7 @@ plugins {
 }
 
 application {
-	mainClassName = 'com.zeroskill.admin-api.AdminApiApplication'
+	mainClassName = 'com.zeroskill.admin_api.AdminApiApplication'
 }
 
 dependencies {
@@ -26,4 +26,12 @@ dependencies {
 	implementation 'javax.xml.bind:jaxb-api:2.4.0-b180830.0359'
 
 	implementation project(':core:common')
+}
+
+springBoot {
+	mainClass = 'com.zeroskill.admin_api.AdminApiApplication'  // 메인 클래스
+}
+
+bootJar {
+	enabled = true  // 실행 가능한 JAR 생성
 }
