+++ api/cms-api/build.gradle
@@ -6,7 +6,7 @@ plugins {
 }
 
 application {
-    mainClassName = 'com.zeroskill.product-api.CmsApiApplication'
+    mainClassName = 'com.zeroskill.cms_api.CmsApiApplication'
 }
 
 dependencies {
@@ -26,5 +26,13 @@ dependencies {
     implementation 'javax.xml.bind:jaxb-api:2.4.0-b180830.0359'
 
     implementation project(':core:common')
+}
+
+springBoot {
+    mainClass = 'com.zeroskill.cms_api.CmsApiApplication'  // 메인 클래스
+}
+
+bootJar {
+    enabled = true  // 실행 가능한 JAR 생성
 }
