+++ api/eureka-api/src/main/java/com/zeroskill/eureka_api/EurekaApiApplication.java
@@ -1,4 +1,4 @@
-package com.zeroskill.eurekaapi;
+package com.zeroskill.eureka_api;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
