+++ api/eureka-api/src/test/java/com/zeroskill/eureka_api/EurekaApiApplicationTests.java
@@ -1,4 +1,4 @@
-package com.zeroskill.eurekaapi;
+package com.zeroskill.eureka_api;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
