+++ api/product-api/build.gradle
@@ -6,7 +6,7 @@ plugins {
 }
 
 application {
-	mainClassName = 'com.zeroskill.product-api.ProductApiApplication'
+	mainClassName = 'com.zeroskill.product_api.ProductApiApplication'
 }
 
 dependencies {
@@ -28,3 +28,10 @@ dependencies {
 	implementation project(':core:common')
 }
 
+springBoot {
+	mainClass = 'com.zeroskill.product_api.ProductApiApplication'  // 메인 클래스
+}
+
+bootJar {
+	enabled = true  // 실행 가능한 JAR 생성
+}
