+++ api/web-bff/build.gradle
@@ -6,11 +6,19 @@ plugins {
 }
 
 application {
-    mainClassName = 'com.zeroskill.web-bff.WebBffApplication'
+    mainClassName = 'com.zeroskill.web_bff.WebBffApplication'
 }
 
 dependencies {
     implementation 'org.springframework.boot:spring-boot-starter-web'
     implementation 'org.springframework.cloud:spring-cloud-starter-openfeign'
     implementation project(':core:common')
+}
+
+springBoot {
+    mainClass = 'com.zeroskill.web_bff.WebBffApplication'  // 메인 클래스
+}
+
+bootJar {
+    enabled = true  // 실행 가능한 JAR 생성
 }
