+++ build.gradle
@@ -4,6 +4,15 @@ plugins {
 	id 'io.spring.dependency-management' version '1.1.6' apply false
 }
 
+allprojects {
+	group = 'com.example'
+	version = '1.0-SNAPSHOT'
+
+	repositories {
+		mavenCentral()
+	}
+}
+
 subprojects {
 	group = 'com.zeroskill'
 	version = '0.0.1-SNAPSHOT'
@@ -12,6 +21,10 @@ subprojects {
 	apply plugin: 'java'
 	apply plugin: 'org.springframework.boot'
 	apply plugin: 'io.spring.dependency-management'
+
+	tasks.named('bootJar') {
+		enabled = false
+	}
 
 	repositories {
 		mavenCentral()
